requirements.txt
alembic==1.3.2
anyjson==0.3.3
apache-airflow==1.10.7
apispec==1.3.3
argcomplete==1.11.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.0
cached-property==1.5.1
cattrs==0.9.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorlog==4.0.2
configparser==3.5.3
confluent-kafka==1.0.0
croniter==0.3.31
cx-Oracle==7.3.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.1.1
docopt==0.6.2
docutils==0.15.2
download==0.3.4
elasticsearch==7.0.4
entrypoints==0.3
Flask==1.1.1
Flask-Admin==1.5.4
Flask-AppBuilder==2.2.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.1
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.2
future==0.16.0
graphviz==0.13.2
gunicorn==19.9.0
hdfs==2.5.8
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
iso8601==0.1.12
itsdangerous==1.1.0
JayDeBeApi==1.1.1
jedi==0.14.0
Jinja2==2.10.1
JPype1==0.7.1
json-merge-patch==0.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kafka-python==1.4.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.21.0
matplotlib==3.1.1
mistune==0.8.4
more-itertools==8.0.2
mysqlclient==1.4.6
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.17.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.0
pendulum==1.4.4
pexpect==4.7.0
pickleshare==0.7.5
prison==0.1.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.7
psycopg2==2.8.4
ptyprocess==0.6.0
py4j==0.10.7
pyarrow==0.16.0
Pygments==2.4.2
PyHive==0.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.2
pyrsistent==0.15.3
pyspark==2.4.4
pyspark-stubs==2.4.0.post7
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.2
pytzdata==2019.3
PyYAML==5.3
pyzmq==18.0.2
qtconsole==4.5.1
redis==2.10.6
redis-py-cluster==1.3.6
requests==2.22.0
Send2Trash==1.5.0
setproctitle==1.1.10
six==1.12.0
SQLAlchemy==1.3.4
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.1
tabulate==0.8.3
tenacity==4.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
thrift==0.11.0
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.2
typing==3.7.4.1
typing-extensions==3.7.4.1
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
WTForms==2.2.1
zipp==0.6.0
zope.deprecation==4.4.0
